Everything works, next step is to improve property modification code
