+++ src/extensibility/dotnet-extensibility/Extensibility/DataSource/DataSourceTypes.cs
@@ -23,13 +23,16 @@ public record DataSourceContext(
 /// <summary>
 /// A read request.
 /// </summary>
+/// <param name="OriginalResourceName">The original resource name.</param>
 /// <param name="CatalogItem">The <paramref name="CatalogItem"/> to be read.</param>
 /// <param name="Data">The data buffer.</param>
 /// <param name="Status">The status buffer. A value of 0x01 ('1') indicates that the corresponding value in the data buffer is valid, otherwise it is treated as <see cref="double.NaN"/>.</param>
 public record ReadRequest(
+    string OriginalResourceName,
     CatalogItem CatalogItem,
     Memory<byte> Data,
-    Memory<byte> Status);
+    Memory<byte> Status
+);
 
 /// <summary>
 /// Reads the requested data.
@@ -55,6 +58,7 @@ internal class ReadRequestManager : IDisposable
     public ReadRequestManager(CatalogItem catalogItem, int elementCount)
     {
         var byteCount = elementCount * catalogItem.Representation.ElementSize;
+        var originalResourceName = catalogItem.Resource.Properties!.GetStringValue(DataModelExtensions.OriginalNameKey)!;
 
         /* data memory */
         var dataOwner = MemoryPool<byte>.Shared.Rent(byteCount);
@@ -68,7 +72,12 @@ internal class ReadRequestManager : IDisposable
         statusMemory.Span.Clear();
         _statusOwner = statusOwner;
 
-        Request = new ReadRequest(catalogItem, dataMemory, statusMemory);
+        Request = new ReadRequest(
+            originalResourceName,
+            catalogItem,
+            dataMemory,
+            statusMemory
+        );
     }
 
     public ReadRequest Request { get; }
